requirements.txt
aiogram==3.0.0b7
Django==4.2
phonenumbers==8.13
django-phonenumber-field==7.1
environs==9.5
